requirements.txt
Flask
Flask-Cors
Flask-Session
Flask-SQLAlchemy
python-dotenv
pymysql
cryptography
